Add tests for the v4->v5 DB upgrade
