+++ rotkehlchen/tests/utils/constants.py
@@ -8,6 +8,7 @@ A_XMR = Asset('XMR')
 A_DASH = Asset('DASH')
 A_IOTA = Asset('IOTA')
 A_BSV = Asset('BSV')
+A_BCH = Asset('BCH')
 A_GBP = Asset('GBP')
 A_SNGLS = Asset('SNGLS')
 A_BNB = Asset('BNB')
